% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner_types.R
\name{create_spinner}
\alias{create_spinner}
\title{Return a list enumerating spinner types and associated css classes}
\usage{
create_spinner(type = .spinner_types[1], ...)
}
\arguments{
\item{type}{The type of spinner to add. You can see the available types of spinners in the package variables .spinner_types}

\item{...}{Further parameters to add the div tag generating the spinners}
}
\description{
Return a list enumerating spinner types and associated css classes
}

