% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinnerTypes.R
\name{createSpinner}
\alias{createSpinner}
\title{Return a list enumerating spinner types and associated css classes}
\usage{
createSpinner(type = .spinner_types[1], color = getOption("spinner.color"))
}
\arguments{
\item{type}{The type of spinner to add. You can see the available types of spinners in the package variables .spinner_types}

\item{color}{The color of the spinner to be applied in HTML}
}
\description{
Return a list enumerating spinner types and associated css classes
}

