\name{shinySpinKit-package}
\alias{shinySpinKit-package}
\alias{shinySpinKit}
\docType{package}
\title{
\packageTitle{shinySpinKit}
}
\description{
\packageDescription{shinySpinKit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{shinySpinKit}
\packageIndices{shinySpinKit}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{shinySpinKit}

Maintainer: \packageMaintainer{shinySpinKit}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
